Redesign actor framework messaging and scheduling
Remove single outstanding message limitation. It's now one per actor.
This masssively simplifies horizons. Previously actors had to manually
report their horizons and the scheduler had to keep track of them.
Now the horizon is simply the actor's next message. The actor is allowed
to replace that message with a new one if an incoming message changes it's
state.
Also, advance was removed (actors can send a message to themselves for
the same effect) and the scheduler loop is massively simplifed.
It just finds the message with the lowest time and dispatches it.

Because actors now own their outgoing message (until it's delivered)
this allows us to optimise the storage.

Actors specify all messages they might send with the outbox macro, which
uses unions and other unsafe code to store the message data inline.